% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{as.day}
\alias{as.day}
\title{Pulls out the day as a number}
\usage{
as.day(x, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{the date}
}
\value{
Returns the day
}
\description{
This extracts a day from a date
}
\examples{
as.day("2014-02-14")
}
