% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{as.month}
\alias{as.month}
\title{Pulls out the month as a number}
\usage{
as.month(x, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{the date}
}
\value{
Returns the month
}
\description{
This extracts a month from a date
}
\examples{
as.month("2014-02-14")
}
