% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{as.year}
\alias{as.year}
\title{Pulls out the year as a number}
\usage{
as.year(x, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{the date}
}
\value{
Returns the year
}
\description{
This extracts a year from a date
}
\examples{
as.year("2014-02-14")
}
