% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{direct}
\alias{direct}
\title{New directory function}
\usage{
direct(x)
}
\arguments{
\item{x}{the suffix of the address}
}
\value{
Returns a full address including OneDrive in full
}
\description{
This is a short cut that adds One Drive into the directory
}
\examples{
direct("lobster/Minor Stuff/MEY")
}
