% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{numchar}
\alias{numchar}
\title{Change factor to number}
\usage{
numchar(x)
}
\arguments{
\item{x}{teh factor}
}
\value{
Returns a number
}
\description{
This Changes a factor to number
}
\examples{
numchar(factor)
}
