% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimeFunc.r
\name{toXL}
\alias{toXL}
\title{Copy to Clipboard}
\usage{
toXL(x, rnames = FALSE)
}
\arguments{
\item{x}{name of data.frame}
}
\value{
Clipboard gets loaded up
}
\description{
Tkes a dataframe and copies to to the clipboard so you can paste it into Excel
}
\examples{
toXL(dat)
}
